+++ GoodMorning/GoodMorning/Priority.swift
@@ -23,4 +23,5 @@ enum Priority {
             return UIColor.blue
         }
     }
+
 }
