+++ GoodMorning/GoodMorning/TODOCollectionViewSection.swift
@@ -1,0 +1,15 @@
+//
+//  TODOCollectionViewSection.swift
+//  GoodMorning
+//
+//  Created by Miro on 2023/08/08.
+//
+
+import Foundation
+
+enum Section: Hashable {
+
+    case main
+
+}
+
